requirements.txt
altgraph==0.17.4
asgiref==3.8.1
bottle==0.13.1
cefpython3==66.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
packaging==24.1
proxy_tools==0.1.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.8.0
python-dotenv==1.0.1
pywebview==5.2
setuptools==75.1.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
